Spectrum_mod: Small change to folding code to adapt to change in numpy masked array behavior.
We'd like the folded spectrum to preserve all the counts of the unfolded, even when the entry is masked. From numpy 1.2 to 1.3, there was a change in behavior. It used to be that adding a number to a masked entry would still change the data for that entry. That seems to no longer be the case. This fix corrects for that.



git-svn-id: https://dadi.googlecode.com/svn/trunk@248 979d6bd5-6d4d-0410-bece-f567c23bd345