Change exception to a warning if data or mask in Spectrum constructor inconsistent data_folded=True.
Primarily, this is movtivated by the fact that many of the tests for arithmetic operations are, e.g. adding a value to every entry of a folded Spectrum. Since
I've decreed that even the masked entries should get this assignment, the 
resulting constructor will fail. I like the behavior of acting on masked values
because it's a check that various oprerations conserve SNP number, so I'm okay
with this just being a warning.



git-svn-id: https://dadi.googlecode.com/svn/trunk@290 979d6bd5-6d4d-0410-bece-f567c23bd345