Improve support for alternate grid definitions.
The extrapolation routines were hard-wired to use the default_grid. To circumvent this, the from_phi routine was modified to store the x-value of the first grid point in fs.extrap_x. The extrapolation routines can now use this, although there is an option to explicitly specify the extrapolation x points.

In order to support this new extrap_x attribute, the Spectrum object also had to be modified to ensure it is copied upon arithmetic.



git-svn-id: https://dadi.googlecode.com/svn/trunk@360 979d6bd5-6d4d-0410-bece-f567c23bd345