Change LRT function to use existing get_godambe code for calculating H and J, to avoid complex duplicated code.
Note that this is a slight change in behavior. The previous code used ll_multinom, so theta was being implicitly tweaked when taking derivatives, rather than being held constant. I believe it is more consistent to hold theta constant, as this code does. In the YRI_CEU example, the difference is tiny. (Adjustment of 0.0230 vs 0.0235.)