Initial support for changing theta in Godambe.
For models with a fixed theta, often theta needs to vary over bootstrap samplings. To support this, added a parameter boot_theta_adjusts, which is used to pass relative values of the theta parameter for each bootstrap data set. Still need to add support to other methods that rely on get_godambe. But those shouldn't be too hard.